Fork wrapper into parent and child pair
Run the wrapper across two processes.
The parent will stay in its original process group, allowing it to get
SIGTSTP and SIGCONT as an un-wrapped QEMU process would.
When the child gets a SIGTSTP from the parent it will SIGSTOP the QEMU
grandchild and then stop itself.  When the child gets a SIGTSTP from any
other process, it will only forward it to the process group that was
originally in the foreground.  This mimics the behavior the terminal
would have produced on Ctrl+Z if the child hadn't put its new process
group into the foreground.