Copy (not hardlink) test-guest binary/ies
When run under `nix build`, the test-guest binary/ies will be built into
the Nix store.  On macOS This is on a separate volume from the temporary
build directory (that hosts `cargo test`s `$CARGO_TARGET_TMPDIR` scratch
directory) so a hardlink cannot be used.

Additionally, QEMU writes to the binary for some reason and the nix
store's copy is read-only.